decomp: SampleData::SizeAs 73.5%→100%, EntityUploader::RockCentralOpComplete 72.7%→98.8%
SampleData: fix switch cases for kVAG/kATRAC/kMP3/kXMA/kNintendoADPCM using
correct integer division expressions and proper int cast before 0x60 addition.

EntityUploader: remove incorrect UploadSecComplete virtual (not in RB3
TourSavable), fix type-2 null guard for UploadComplete call.

+++ src/system/synth/SampleData.cpp
@@ -120,7 +120,7 @@ int SampleData::SizeAs(Format fmt) const {
         return mNumSamples / 5;
     }
     case kNintendoADPCM: {
-        return ((mNumSamples * 2) / 3.4f) + 0x60;
+        return (int)((float)(mNumSamples * 2) / 3.4f) + 0x60;
     }
     default: {
         MILO_ASSERT(0, 0x136);
